requirements.txt
numpy
loguru
ete3
six
prompt_toolkit
prettytable
python-dateutil
PyPDF2
matplotlib
pandas
yahoo_fin
yfinance
typing-extensions
fpdf
requests
legacy-cgi # if using Python 3.13 or greater
loguru
plotly
scikit-learn
plaid-python>=16.0.0
cryptography>=41.0.0

